cloud_functions/scrape_weather/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
requests==2.*
